Moved all query-hooks to separate folder.
