+++ src/components/Providers/Providers.js
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types';
 
 import { MenuContext, GlobalContext } from '@contexts';
 import { StaticImagesService } from '@services';
-import { useIsOpened, useStaticImages } from '@hooks';
+import { useIsOpened } from '@hooks';
+import { useStaticImages } from '@hooks/queries';
 
 const Providers = ({ children }) => {
   const menuState = useIsOpened();
