+++ src/hooks/index.js
@@ -1,15 +1,4 @@
-import useDribbbleShots from './useDribbbleShots';
 import useGetBreakpoint from './useGetBreakpoint';
 import useIsOpened from './useIsOpened';
-import useStaticImages from './useStaticImages';
-import useSiteMetadata from './useSiteMetadata';
-import useSitePages from './useSitePages';
 
-export {
-  useDribbbleShots,
-  useGetBreakpoint,
-  useIsOpened,
-  useStaticImages,
-  useSiteMetadata,
-  useSitePages,
-};
+export { useGetBreakpoint, useIsOpened };
