+++ src/hooks/queries/index.js
@@ -1,0 +1,6 @@
+import useDribbbleShots from './useDribbbleShots';
+import useStaticImages from './useStaticImages';
+import useSiteMetadata from './useSiteMetadata';
+import useSitePages from './useSitePages';
+
+export { useDribbbleShots, useStaticImages, useSiteMetadata, useSitePages };
