+++ src/hooks/queries/useStaticImages.js
